Restructured data objects
Entities to be utilized by application.
Restructured Agent model
